Added cast to make SkIntToScalar happy with size_t in debug (in SkDrawCommand.cpp)
git-svn-id: http://skia.googlecode.com/svn/trunk@7495 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ debugger/SkDrawCommand.cpp
@@ -329,7 +329,7 @@ DrawPoints::DrawPoints(SkCanvas::PointMode mode, size_t count,
     this->fDrawType = DRAW_POINTS;
 
     this->fInfo.push(SkObjectParser::PointsToString(pts, count));
-    this->fInfo.push(SkObjectParser::ScalarToString(SkIntToScalar(count),
+    this->fInfo.push(SkObjectParser::ScalarToString(SkIntToScalar((unsigned int)count),
                                                     "Points: "));
     this->fInfo.push(SkObjectParser::PointModeToString(mode));
     this->fInfo.push(SkObjectParser::PaintToString(paint));
